+++ .vscode/typescript.json.code-snippets
@@ -12,8 +12,13 @@
   "Bot Interaction": {
     "prefix": ["Interaction"],
     "body": [
+      "import { ApplicationCommandOptionData, CommandInteraction } from 'discord.js';",
       "import { InteractionCommand } from '../core/interactions/interactionCommand';\n",
-      "export class ${1: className} extends InteractionCommand {\t async execute(): Promise<void>  {",
+      "export class ${1: className} extends InteractionCommand {",
+      "name = '';",
+      "description = '';",
+      "options: ApplicationCommandOptionData[] = [];\n",
+      "async execute(interaction: CommandInteraction): Promise<void>  {",
       "\t$0",
       "}}"
     ],
